+++ docs/conf.py
@@ -247,3 +247,5 @@ texinfo_documents = [
 # Example configuration for intersphinx: refer to the Python standard library.
 intersphinx_mapping = {'http://docs.python.org/': None}
 
+# autodoc configuration
+autoclass_content = 'both'